Figure out how to use serde_macro and friends correctly
After a lot of messing around, I managed to figure out how to use Serde 0.4.0
correctly. I *really* wish I'd found
http://erickt.github.io/blog/2015/05/18/serde-0-dot-4-0-now-supports-macros-in-stable-rust/
earlier, as it actually has the process documented, where the GitHub README.md
does not.

At the moment, all this really has it a bogus Point struct which I can
serialize to a string and deserialize from a string as JSON, and prove to my
satisfaction that it's valid both ways. Next I have to dig into the Avro schema
spec and turn that into a viable object model I can manipulate with Serde.