Move to ravro::schema::Schema approach
So at the moment I'm having a *lot* of problem groking how to get polymorphism
to work correctly in Rust. So rather than continuing to beat my head against
the wall, I'm simply goint to take another approach. Namely, since at the end
of the day, all the various ways of parsing the scheam will result in the same
object model at the end of the day, so just work with it. The downside is that
currently serializing that model to JSON produces way more text than is really
necessary. For example:

{"type":"null","namespace":null,"doc":null,"aliases":[],"fields":[]}

instead of the more cannonical (although not the strict Avro cannonical form):

{"type":"null"}

For now, I'm going to accept the drawback and see what I can do later.