Flesh out implementation of schema::from_str and test it
Obviously the majority of the time we want people to deserialize Avro schemas
from strings/files/whatever rather than having to construct objects by hand.
Hence the need to flesh out the from_str function to start which can cope with
"raw" primitive types and JSON objects and arrays.

Right now, it really can cope only with the raw primitive types and the
primitive types as a JSON object.