Test with ravro::schema::Schema rather than deserialized JSON
So the deserialization tests really should be testing against a hand crafted
oracle (i.e., a hand build Schema object) rather than relying on Serde to
correctly deserialize the JSON I was handing to it.  That's what I'm supposed
to be testing (along one branch, anyway) with the schema::from_str() function.