Use str instead of String where possible
In most of the places where I've been using String, it's not necessary; the
strings are fixed sized because they are either parsed from JSON or known in
advance.  Unfortunately, Serde doesn't seem to know what to do with lifetime
markers yet, and it produces compilation errors, so things like

struct Blah<'a> { foo: &'a str, }

won't compile cleanly because the Serde code generation treats 'a as a regular
type rather than as a lifetime marker to effectively be ignored, or as a
PhantomData type perhaps. I don't know enough yet to comment on a solution.

The consequence is that the natural change to

pub struct Schema<'a> {
    #[serde(rename="type")]
    type_name: &'a str;
    // ...
}

can't be done. I have to stick with type_name being a String for now.