Add schema::new_rec() function and re-org tests a bit
The big lesson learned during this change was that the following line to return
a schema value

    Ok(schema);

is not only wrong (don't need the semi-colon, the usual mistake we Rust newbies
make) but when coupled with the Return<T,E> and Ok/Err setup, the error you get
out of the Rust compiler is COMPLETELY NOT WHAT YOU EXPECT (it complains about
not being able to infer the type of '_', which is not helpful at all). So, note
to self: Always always always check for extra semicolons at the end of your
functions.