Starting in on enum support, added Schema::new_enum()
The immediate sticker is the fact that if I add a symbols member to the Schema
struct, it's going to get serialized when I don't want it to. Namely, with
records, maps, etc.

The Serde JSON serialization support does have some definite limitations in
comparison with Jackson. Although to be fair, I'm betting Jackson has been
around a lot longer.