Start new Schema implementation
Staring new approach for implementing Schema, by *not* using Serde to do the
serialization and deserialization on the objects directly.  Instead, I'm going
to acknowledge I'm using JSON and handle serialization with Serde, but through
the serde::json::* interface rather than code generation on my structs/enums. I
think it's going to be a lot cleaner in the end. #shrug. We'll see.