Replace str.to_string() with String::from(str) in tests
Now that I've discovered String::from(str), which is potentially faster than
"blah".to_string() (apparently has a cleaner code path), I'm using it as a
matter of personal prefernce. Not only for the theoretical (as far as I'm
concerned) perf improvement, I simply prefer the visual flow of writing a str
to String conversion this way.