Start working on record support
Implemented the core::convert::From trait to convert a Schema object to a String.

Starting to support the record complex type of the Schema::Object type. Added a
Schema::is_record() helper method, and with the From trait support we have
basic serialization; under the covers we are using the Serde JSON
to_string_pretty() function after converting the Schema instance we are passed
to a Schema::Object.