Improve Schema::fullname() impl and add Schema::valid_fullname()
Improved the implementation of Schema:::fullname() to meet the specification
requirements for records, enums, and fixed deinfitions at the top level. It
doesn't take into account field names and scoping rules yet.

Also added a Schema::valid_fullname() method to determine if the fullname of
the record, etc. element is actually valid according to the Avro specification.

Note that I'm not convinced that this is the best way to implement this, and I
might go back to returning a Result<String, Error> from fullname() rather than
a bare String. That would eliminate the need for valid_fullname() and make it
unambiguous that the Schema has problems.