Changed signature of Schmea::fullname(), removed Schema::valid_fullname()
To make life simpler in the long run, and to keep things more in the Rust
style, I changed the signature of fullname() to return Result<String, &'static
str> and made valid_fullname() use the same return signature, and also made it
private. So now it's impossible to use an invalid fullname() from a poorly
constructed Schema instance.