Removing old code, first pass
I'm pretty much committed to the "new style" for the Schema enum, so I'm going
finally get rid of the old code that I was hesitant to toss out. This is the
first pass, the second pass will be to get rid of the lib.rs.in file, as I'm no
longer doing any code generation.