Re-org Schema model code, remove now unnecessary code generation
Now that I'm no longer doing any Serde code generation, remove all of the build
goo that I had in place for that. I'm still using the Serde JSON support, so my
reliance on the core Serde library is there, but all the rest of it is now
gone. I may discover a need for the Serde code generation in the future, but at
the moment it's not there.

Additionally, the Schema model code has been moved to src/schema/model.rs from
src/lib.rs.in, which is more intuitive. The ravro::schema::model module itself
is private, the public elements within it are publicly re-exported via
ravro::schema::*.