Added support for the array schema complex type
Added ArrayBuilder with the ability to add the items attribute, and that's it.
Needed to update the Schema::is_array() method to differentiate between
Schema::Array and Schema::Object with a type value of "array".

I'm really thinking that I'd be better off renaming Schema::Array to
Schema::Union, because that's really what it is, despite the fact that it's
represented as a JSON array and it's initially described as an array.