Renamed Schema::Array to Schema::Union
Renaming Schema::Array to Schema::Union, which more closely matches the
specification in meaning and intent, has some obvious knock on effects. The
obvious being there is now is Schema::is_union() method and the
Schema::is_array() is purely for determining if a Schema::Object represents the
array complex type.