Added UnionBuilder, updated builders to take str's instead of Strings
Added a UnionBuilder to make it easier to create Schema::Union instances. I
also changed the signature of all the builder methods that previously took
Strings to take &str instead, as it makes the call site a lot cleaner, and the
memory use is the same at the end of the day.