Start implementing Schema::is_valid() to ensure schemas are correct according to the specification
The Builders will help to manually construct Schema instances correctly, but
you can still mess it up. And we have to parse arbitrary JSON into Schema
instances and ensure they are valid. To do so, we need to enforce a bunch of
rules required by the specification.

Starting simple by implementing is_valid() for the Schema::String type with
some basic unit tests.