Partial implementation of Schema::is_valid() for records
Records are probably the most complex things to validate, due to fields,
because each field has a type and an optional default value.  If it has a
default, the default value JSON type has to match the field Avro type according
to the match table provided in the specification.

So far, I've implemented field type and default type comparison for simple
string types. Need to move on to objects and unions.