Add support for checking default type matches for unions in record field types
Unions in record field types aren't too bad, because the rule is the default
value applies only to the first element of the union.

Note that the other rules about what types are allowed in a union still need to
be applied (i.e., only one of each type, unless they are named types, etc.).